added predicting linear regression function.
Now we just need to test all of the methods we created. We can do it during class. Before then I will try to get 100% coverage on my branch and you can do it on your branch and then if we both get 100% then we should be set for the testing portion for our code